+++ packages/backend/src/nest/socket/socket.service.ts
@@ -172,9 +172,6 @@ export class SocketService extends EventEmitter implements OnModuleInit {
           ownerCertificate: payload.certificate,
           rootCa: payload.permsData.certificate,
         }
-
-        console.log('Metadata from state-manager', communityMetadataPayload)
-
         this.emit(SocketActionTypes.SEND_COMMUNITY_METADATA, communityMetadataPayload)
       })
 
@@ -214,7 +211,7 @@ export class SocketService extends EventEmitter implements OnModuleInit {
   public listen = async (port = this.configOptions.socketIOPort): Promise<void> => {
     return await new Promise(resolve => {
       if (this.serverIoProvider.server.listening) resolve()
-      this.serverIoProvider.server.listen(this.configOptions.socketIOPort, () => {
+      this.serverIoProvider.server.listen(this.configOptions.socketIOPort, '127.0.0.1', () => {
         this.logger(`Data server running on port ${this.configOptions.socketIOPort}`)
         resolve()
       })