+++ packages/e2e-tests/src/utils.ts
@@ -1,5 +1,5 @@
 import { Browser, Builder, type ThenableWebDriver } from 'selenium-webdriver'
-import { spawn, exec, type ChildProcessWithoutNullStreams, execSync } from 'child_process'
+import { spawn, exec, execSync, type ChildProcessWithoutNullStreams } from 'child_process'
 import { type SupportedPlatformDesktop } from '@quiet/types'
 import getPort from 'get-port'
 import path from 'path'
@@ -15,8 +15,6 @@ export interface BuildSetupInit {
 
 export class BuildSetup {
   private driver?: ThenableWebDriver | null
-  public containerId?: string
-  public ipAddress?: string
   public port?: number
   public debugPort?: number
   public dataDir?: string
@@ -104,16 +102,16 @@ export class BuildSetup {
     await this.initPorts()
     const env = {
       DATA_DIR: this.dataDir || 'Quiet',
-      DEBUG: 'backend*',
+      DEBUG: 'backend*,desktop*',
     }
     if (process.platform === 'win32') {
       console.log('!WINDOWS!')
-      this.child = spawn(`cd node_modules/.bin & chromedriver.cmd --port=${this.port}`, [], {
+      this.child = spawn(`cd node_modules/.bin & chromedriver.cmd --port=${this.port} --verbose`, [], {
         shell: true,
         env: Object.assign(process.env, env),
       })
     } else {
-      this.child = spawn(`node_modules/.bin/chromedriver --port=${this.port}`, [], {
+      this.child = spawn(`node_modules/.bin/chromedriver --port=${this.port} --verbose`, [], {
         shell: true,
         detached: false,
         env: Object.assign(process.env, env),
@@ -127,7 +125,7 @@ export class BuildSetup {
     )
 
     this.child.on('error', () => {
-      console.log('ERROR')
+      console.error('ERROR')
       this.killNine()
     })
 
@@ -145,7 +143,7 @@ export class BuildSetup {
       console.log('message', data)
     })
     this.child.on('error', data => {
-      console.log('error', data)
+      console.error('error', data)
     })
 
     this.child.stdout.on('data', data => {
@@ -153,11 +151,17 @@ export class BuildSetup {
     })
 
     this.child.stderr.on('data', data => {
-      console.error(`stderr: ${data}`)
+      // Quiet logs (handled by 'debug' package) are available in stderr and only with 'verbose' flag on chromedriver
+      const trashLogs = ['DevTools', 'COMMAND', 'INFO:CONSOLE', '[INFO]:', 'libnotify-WARNING', 'ALSA lib']
+      const dataString = `${data}`
+      for (const l of trashLogs) {
+        if (dataString.includes(l)) return
+      }
+      console.log(dataString)
     })
 
     this.child.stdin.on('data', data => {
-      console.error(`stdin: ${data}`)
+      console.log(`stdin: ${data}`)
     })
   }
 
@@ -196,7 +200,7 @@ export class BuildSetup {
       }
     }
     if (this.driver == null || this.driver === undefined) {
-      throw new Error('elo')
+      throw new Error('No driver')
     }
 
     return this.driver