+++ packages/mobile/src/screens/PossibleImpersonationAttack/PossibleImpersonationAttack.types.ts
@@ -1,0 +1,10 @@
+import { PossibleImpersonationAttackRouteProps } from '../../route.params'
+
+export interface PossibleImpersonationAttackScreenProps {
+  route: PossibleImpersonationAttackRouteProps
+}
+export interface PossibleImpersonationAttackComponentProps {
+  handleBackButton: () => void
+  communityName: string
+  leaveCommunity: () => void
+}